fix a compile error in Mac.

+++ src/c_api_common.cc
@@ -24,8 +24,8 @@ DLManagedTensor* CreateTmpDLManagedTensor(const DGLArgValue& arg) {
 
 PackedFunc ConvertNDArrayVectorToPackedFunc(const std::vector<NDArray>& vec) {
     auto body = [vec](DGLArgs args, DGLRetValue* rv) {
-        const size_t which = args[0];
-        if (which >= vec.size()) {
+        const int which = args[0];
+        if (which >= (int) vec.size()) {
             LOG(FATAL) << "invalid choice";
         } else {
             *rv = std::move(vec[which]);
